Port `class` snippet
